rename and distribute routes among router files
