+++ views/partials/UserBar.hbs
@@ -28,10 +28,10 @@
       </li>
       {{else}}
       <li class="nav-item">
-        <a class="nav-link" href="/sign-up">Register</a>
+        <a class="nav-link" href="/user/sign-up">Register</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="/sign-in">Log In</a>
+        <a class="nav-link" href="/user/sign-in">Log In</a>
       </li>
       {{/if}}
     </ul>
