+++ views/sign-in.hbs
@@ -1,5 +1,5 @@
 <div class="container">
-  <form class="form-signin" action="/sign-in" method="POST">
+  <form class="form-signin" action="/user/sign-in" method="POST">
     <h2 class="form-signin-heading">Please sign in</h2>
 
     {{#if invalidLogin}}
