+++ views/sign-up.hbs
@@ -1,5 +1,5 @@
 <div class="container">
-  <form class="form-signup" action="/sign-up" method="POST">
+  <form class="form-signup" action="/user/sign-up" method="POST">
     <h2 class="form-signup-heading">Sign Up</h2>
 
  <div class="form-group">
